Recargar Saldo Máquina

+++ DAO/recargaDAO.cs
@@ -52,8 +52,8 @@ namespace DAO {
 
             return codigo;
         }
-
-        public string actualizarSaldo(double monto, string dni) {
+        
+        public string actualizarSaldo(decimal monto, string dni) {
             string m = "";
             cn.getcn.Open();
