+++ Entidades/RecargaEntity.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace Entidades {
     public class RecargaEntity {
         public String codigo { get; set; }
-        public double monto { get; set; }
+        public decimal monto { get; set; }
         public String fechaReg { get; set; }
         public String dni { get; set; }
         public String usuario { get; set; }
